Coded the interceptors from #14 and #13
