--- src/app/core/interceptors/post/post.interceptor.spec.ts
@@ -1,16 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { PostInterceptor } from './post.interceptor';
-
-describe('PostInterceptor', () => {
-  beforeEach(() => TestBed.configureTestingModule({
-    providers: [
-      PostInterceptor
-      ]
-  }));
-
-  it('should be created', () => {
-    const interceptor: PostInterceptor = TestBed.inject(PostInterceptor);
-    expect(interceptor).toBeTruthy();
-  });
-});
